library-ify kona host
