+++ bin/host/Cargo.toml
@@ -26,10 +26,10 @@ kona-mpt = { path = "../../crates/mpt", version = "0.0.2" }
 kona-derive = { path = "../../crates/derive", version = "0.0.2", features = ["online"] }
 
 # external
-alloy-provider = { version = "0.1" } 
-alloy-transport-http = { version = "0.1" } 
-alloy-rpc-client = { version = "0.1" } 
-alloy-rpc-types = { version = "0.1" } 
+alloy-provider = { version = "0.1" }
+alloy-transport-http = { version = "0.1" }
+alloy-rpc-client = { version = "0.1" }
+alloy-rpc-types = { version = "0.1" }
 reqwest = "0.12"
 tokio = { version = "1.37.0", features = ["full"] }
 futures = "0.3"
